MAINT: Refactor models for ABC. Default isotropic anisotropy constants in stiffness tensor generation
