+++ convergence_analysis/convergence_analysis_models/model_convergence_ABC_heterogeneity.py
@@ -12,105 +12,12 @@ import porepy as pp
 
 sys.path.append("../../")
 
-from models import DynamicMomentumBalanceABCLinear
+from .model_convergence_ABC import ABCModel
 
 import sympy as sym
 
 
-class BoundaryConditionsUnitTest:
-    def bc_type_mechanics(self, sd: pp.Grid) -> pp.BoundaryConditionVectorial:
-        """Method for assigning boundary condition type.
-
-        Assigns the following boundary condition types:
-            * North and south: Roller boundary (Neumann in x-direction and Dirichlet in
-              y-direction)
-            * West: Dirichlet
-            * East: Robin
-
-        Parameters:
-            sd: The subdomain whose bc type is assigned.
-
-        Return:
-            The boundary condition object.
-
-        """
-        # Fetch boundary sides and assign type of boundary condition for the different
-        # sides
-        bounds = self.domain_boundary_sides(sd)
-        bc = pp.BoundaryConditionVectorial(sd, bounds.all_bf, "dir")
-
-        # East side: Absorbing
-        bc.is_dir[:, bounds.east] = False
-        bc.is_rob[:, bounds.east] = True
-
-        # North side: Roller
-        bc.is_dir[0, bounds.north + bounds.south] = False
-        bc.is_neu[0, bounds.north + bounds.south] = True
-
-        # Calling helper function for assigning the Robin weight
-        self.assign_robin_weight(sd=sd, bc=bc)
-        return bc
-
-    def bc_values_stress(self, boundary_grid: pp.BoundaryGrid) -> np.ndarray:
-        """Method for assigning Neumann and Robin boundary condition values.
-
-        Specifically, for the Robin values, this method assigns the values corresponding
-        to absorbing boundary conditions with a second order approximation of u_t in:
-
-            sigma * n + alpha * u_t = G
-
-        Robin/Absorbing boundaries are employed for the east boundary. Zero Neumann
-        conditions are assigned for the north and south boundary.
-
-        Parameters:
-            boundary_grid: The boundary grids on which to define boundary conditions.
-
-        Returns:
-            Array of boundary values.
-
-        """
-        if boundary_grid.dim != (self.nd - 1):
-            return np.array([])
-
-        data = self.mdg.boundary_grid_data(boundary_grid)
-        sd = boundary_grid.parent
-        boundary_faces = sd.get_boundary_faces()
-        name = "boundary_displacement_values"
-
-        if self.time_manager.time_index == 0:
-            return self.initial_condition_bc(boundary_grid)
-
-        displacement_values_0 = pp.get_solution_values(
-            name=name, data=data, time_step_index=0
-        )
-        displacement_values_1 = pp.get_solution_values(
-            name=name, data=data, time_step_index=1
-        )
-
-        # According to the expression for the absorbing boundaries we have a coefficient
-        # 2 in front of the values u_(n-1) and -0.5 in front of u_(n-2):
-        displacement_values = 2 * displacement_values_0 - 0.5 * displacement_values_1
-
-        # Transposing and reshaping displacement values to prepare for broadcasting
-        displacement_values = displacement_values.reshape(
-            boundary_grid.num_cells, self.nd, 1
-        )
-
-        # Assembling the vector representing the RHS of the Robin conditions
-        total_coefficient_matrix = self.total_coefficient_matrix(sd=sd)
-        robin_rhs = np.matmul(total_coefficient_matrix, displacement_values).squeeze(-1)
-        robin_rhs *= sd.face_areas[boundary_faces][:, None]
-        robin_rhs = robin_rhs.T
-
-        boundary_sides = self.domain_boundary_sides(sd)
-        for direction in ["north", "south"]:
-            inds = np.where(getattr(boundary_sides, direction))[0]
-            inds = np.where(np.isin(boundary_faces, inds))[0]
-            robin_rhs[:, inds] *= 0
-
-
-        return robin_rhs.ravel("F")
-
+class BoundaryConditions:
     def bc_values_displacement(self, bg: pp.BoundaryGrid) -> np.ndarray:
         """Method for setting Dirichlet boundary values.
 
@@ -137,37 +44,6 @@ class BoundaryConditionsUnitTest:
         )
 
         return values.ravel("F")
-
-    def initial_condition_bc(self, bg: pp.BoundaryGrid) -> np.ndarray:
-        """Method for setting initial boundary values for 0th and -1st time step.
-
-        Parameters:
-            bg: Boundary grid whose boundary displacement value is to be set.
-
-        Returns:
-            An array with the initial displacement boundary values.
-
-        """
-        dt = self.time_manager.dt
-        vals_0 = self.initial_condition_value_function(bg=bg, t=0)
-        vals_1 = self.initial_condition_value_function(bg=bg, t=0 - dt)
-
-        data = self.mdg.boundary_grid_data(bg)
-
-        # The values for the 0th and -1th time step are to be stored
-        pp.set_solution_values(
-            name="boundary_displacement_values",
-            values=vals_0,
-            data=data,
-            time_step_index=0,
-        )
-        pp.set_solution_values(
-            name="boundary_displacement_values",
-            values=vals_1,
-            data=data,
-            time_step_index=1,
-        )
-        return vals_0
 
     def initial_condition_value_function(
         self, bg: pp.BoundaryGrid, t: float
@@ -453,10 +329,10 @@ class ExactHeterogeneousSigmaAndForce:
         return full_force_array
 
 
-class ABCModel(
-    BoundaryConditionsUnitTest,
+class ABCModelHeterogeneous(
+    BoundaryConditions,
     ConstitutiveLawsAndSource,
     InitialConditions,
     ExactHeterogeneousSigmaAndForce,
-    DynamicMomentumBalanceABCLinear,
+    ABCModel,
 ): ...
