+++ utils/anisotropy_mixins.py
@@ -8,6 +8,7 @@ from utils.utility_functions import (
 )
 from utils.stiffness_tensors import TensorAllowingForCustomFields
 
+
 class TransverselyIsotropicTensorMixin:
     def stiffness_tensor(self, subdomain: pp.Grid):
         """Compute the stiffness tensor for a given subdomain."""
@@ -16,6 +17,11 @@ class TransverselyIsotropicTensorMixin:
             self=self,
             sd=subdomain,
         )
+
+        try:
+            h = self.heterogeneity_factor
+        except AttributeError:
+            h = 1
 
         # Preparing basis arrays for inner and outer domains
         inner = np.zeros(subdomain.num_cells)
@@ -42,18 +48,27 @@ class TransverselyIsotropicTensorMixin:
         mu_orthogonal_mat = stiffness_matrices["mu_perpendicular"]
 
         # Extract individual constants from the anisotropy constants dictionary
-        anisotropy_constants = self.params["anisotropy_constants"]
+        anisotropy_constants = self.params.get(
+            "anisotropy_constants",
+            {
+                "mu_parallel": self.solid.shear_modulus,
+                "mu_orthogonal": self.solid.shear_modulus,
+                "lambda_parallel": 0.0,
+                "lambda_orthogonal": 0.0,
+                "volumetric_compr_lambda": self.solid.lame_lambda,
+            },
+        )
 
-        volumetric_compr_lambda = anisotropy_constants["volumetric_compr_lambda"]
-        lambda_parallel = anisotropy_constants["lambda_parallel"]
-        lambda_orthogonal = anisotropy_constants["lambda_orthogonal"]
-        mu_parallel = anisotropy_constants["mu_parallel"]
-        mu_orthogonal = anisotropy_constants["mu_orthogonal"]
+        volumetric_compr_lambda = anisotropy_constants["volumetric_compr_lambda"] * h
+        lambda_parallel = anisotropy_constants["lambda_parallel"] * h
+        lambda_orthogonal = anisotropy_constants["lambda_orthogonal"] * h
+        mu_parallel = anisotropy_constants["mu_parallel"] * h
+        mu_orthogonal = anisotropy_constants["mu_orthogonal"] * h
 
         # Standard material values: assigned to the outer domain
         lmbda = self.solid.lame_lambda * outer
         mu = self.solid.shear_modulus * outer
-        
+
         # Values for inner domain with anisotropic constants
         mu_parallel_inner = mu_parallel * inner
         mu_orthogonal_inner = mu_orthogonal * inner
